* rename fail2ban-ssh to repeat-ssh
The name "fail2ban-ssh" causes a conflit between the rule
"fail2ban-fail2ban-ssh" (created by the script)
and "fail2ban-sshd" (created by fail2ban for the rule sshd)

Fail2ban internally suppreses the prefix "fail2ban".
In result, there is the rule "sshd" and "fail2ban-ssh".
But, when checking if a rule exists, the full name is used.
Dans the token "fail2ban-ssh" exists in "fail2ban-sshd", so
a rule is not applied.